core: Extract integrator communication
