+++ src/core/integrate.hpp
@@ -113,6 +113,24 @@ int integrate(int n_steps, int reuse_forces);
  */
 int python_integrate(int n_steps, bool recalc_forces, bool reuse_forces);
 
+/** Start integrator.
+ *  @param n_steps       how many steps to do.
+ *  @param reuse_forces  whether to trust the old forces for the first half step
+ *  @return nonzero on error
+ */
+int mpi_integrate(int n_steps, int reuse_forces);
+
+/** Steepest descent main integration loop
+ *
+ *  Integration stops when the maximal force is lower than the user limit
+ *  @ref SteepestDescentParameters::f_max "f_max" or when the maximal number
+ *  of steps @p steps is reached.
+ *
+ *  @param steps Maximal number of integration steps
+ *  @return number of integrated steps
+ */
+int mpi_steepest_descent(int steps);
+
 /** @brief Set the steepest descent integrator for energy minimization.
  *  @retval ES_OK on success
  *  @retval ES_ERROR on error
