+++ src/python/espressomd/integrate.pxd
@@ -27,6 +27,7 @@ cdef extern from "config.hpp":
 
 cdef extern from "integrate.hpp" nogil:
     cdef int python_integrate(int n_steps, cbool recalc_forces, int reuse_forces)
+    cdef int mpi_steepest_descent(int max_steps)
     cdef void integrate_set_sd()
     cdef void integrate_set_nvt()
     cdef int integrate_set_steepest_descent(const double f_max, const double gamma,
@@ -66,6 +67,3 @@ IF(STOKESIAN_DYNAMICS or STOKESIAN_DYNAMICS_GPU):
 cdef inline int _integrate(int nSteps, cbool recalc_forces, int reuse_forces):
     with nogil:
         return python_integrate(nSteps, recalc_forces, reuse_forces)
-
-cdef extern from "communication.hpp":
-    int mpi_steepest_descent(int max_steps)
